+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 **Fixed**
 
 * Fix Sabaki crashing when pressing enter in autoplay mode
+* Fix 'Clear Annotations' not working
 
 ## [Sabaki v0.31.1][v0.31.1] (2017-06-24)